+++ Runtime/StateMachines/IStateTransition.cs
@@ -1,9 +1,0 @@
-﻿namespace DapperDino.DapperTools.StateMachines
-{
-    public interface IStateTransition
-    {
-        State NextState { get; }
-
-        bool ShouldTransition();
-    }
-}