+++ Runtime/StateMachines/StateMachineBehaviour.cs
@@ -7,8 +7,18 @@ namespace DapperDino.DapperTools.StateMachines
         [SerializeField] private State startingState = null;
 
         private StateMachine stateMachine;
+        private StateMachine StateMachine
+        {
+            get
+            {
+                if (stateMachine != null) { return stateMachine; }
+                stateMachine = new StateMachine(startingState);
+                return stateMachine;
+            }
+        }
 
-        private void Start() => stateMachine = new StateMachine(startingState);
-        private void Update() => stateMachine.Tick();
+        private void Update() => StateMachine.Tick();
+
+        public void ChangeState(State state) => StateMachine.ChangeState(state);
     }
 }