+++ Runtime/StateMachines/StateTransition.cs
@@ -1,13 +1,28 @@
-﻿using UnityEngine;
+﻿using System;
+using System.Collections.Generic;
+using UnityEngine;
 
 namespace DapperDino.DapperTools.StateMachines
 {
-    public abstract class StateTransition : MonoBehaviour, IStateTransition
+    [Serializable]
+    public class StateTransition
     {
         [SerializeField] private State nextState = null;
+        [SerializeField] private List<StateTransitionCondition> conditions = new List<StateTransitionCondition>();
 
         public State NextState => nextState;
 
-        public abstract bool ShouldTransition();
+        public bool ShouldTransition()
+        {
+            foreach (var condition in conditions)
+            {
+                if (!condition.IsMet())
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
     }
 }