+++ Runtime/StateMachines/StateTransitionCondition.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace DapperDino.DapperTools.StateMachines
+{
+    public abstract class StateTransitionCondition : MonoBehaviour
+    {
+        public abstract bool IsMet();
+    }
+}